+++ src/instrument/ctf/InstrumentTaskId.cpp
@@ -1,5 +1,5 @@
 #include "InstrumentTaskId.hpp"
 
 namespace Instrument {
-	std::atomic<size_t> _nextTaskId(1);
+	std::atomic<uint32_t> _nextTaskId(1);
 }
